dev revamp

+++ server/rest/groups.py
@@ -3,24 +3,31 @@ from flask import request, abort, make_response, session
 from flask_restful import Resource
 
 from config import db
-from models import TaskUpdate
+from models import Group
 
 
-class TaskUpdates(Resource):
+class Groups(Resource):
     def get(self):
         # user_id = session.get('user_id')
         # if not user_id:
         #     return make_response({"error": "User not logged in"}, 401)
-        updates = [update.to_dict(rules=('-task',)) for update in TaskUpdate.query.all()]
-        return make_response(updates, 200)
+        groups_query = Group.query
+        
+        project_filter = request.args.get("project_id")
+        if project_filter != None:
+            groups_query = groups_query.filter(Group.project_id == project_filter)
+        else:
+            return make_response({"error": "Group must be specified"}, 422)
+        
+        groups = [group.to_dict(rules=('-master_tasks','-project')) for group in groups_query.all()]
+        return make_response(groups, 200)
     
     def post(self):
         data = request.get_json()
         try:
-            new_record = TaskUpdate(
-                task_id = data['task_id'],
-                task_status = data['task_status'],
-                message = data['message'],
+            new_record = Group(
+                name = data['name'],
+                project_id = data['project_id']
             )
         except ValueError as e:
             abort(422, e.args[0])
@@ -31,10 +38,10 @@ class TaskUpdates(Resource):
         return make_response(new_record.to_dict(), 201)
     
 
-class TaskUpdateByID(Resource):
+class GroupByID(Resource):
     @classmethod
     def find_model_by_id(cls, id):
-        model = TaskUpdate.query.filter_by(id=id).first()
+        model = Group.query.filter_by(id=id).first()
         if model:
             return model
         else:
@@ -44,7 +51,7 @@ class TaskUpdateByID(Resource):
         model = self.__class__.find_model_by_id(id)
         if not model:
             return make_response({"error": f"Model ID: {id} not found"}, 404)
-        return make_response(model.to_dict(rules=('-task',)), 200)
+        return make_response(model.to_dict(rules=('-master_tasks','-project')), 200)
            
     def patch(self, id):
         model = self.__class__.find_model_by_id(id)
@@ -66,6 +73,14 @@ class TaskUpdateByID(Resource):
         if not model:
             return make_response({"error": f"Model ID: {id} not found"}, 404)
         
+        # if there are tasks linked to this group, do not allow delete
+        if model.master_tasks:
+            tasks_list = ""
+            for task in model.master_tasks:
+                tasks_list = tasks_list + f'{task.name} (ID: {task.id}), '
+            return make_response({"error": f"There are tasks linked to this group: {tasks_list}"}, 424) # 424: failed dependency
+        
         db.session.delete(model)
         db.session.commit()
         return make_response("", 204)
+
